+++ packages/app/src/services/__tests__/settings-database.test.ts
@@ -1,14 +1,14 @@
+import type * as SQLiteTypes from 'expo-sqlite';
+
 import {
   _resetForTesting,
-  initSettingsDatabase,
   getSettings,
-  updateTheme,
+  initSettingsDatabase,
+  isOnboarded,
   updateLanguage,
   updateOnboarded,
-  isOnboarded,
-} from '../settings-database';
-
-import type * as SQLiteTypes from 'expo-sqlite';
+  updateTheme,
+} from '../database/settings-database';
 
 const mockExecAsync = jest.fn().mockResolvedValue(undefined);
 const mockRunAsync = jest.fn().mockResolvedValue(undefined);
@@ -139,10 +139,9 @@ describe('settings-database', () => {
     it('runs UPDATE with the given theme', async () => {
       await updateTheme('dark');
 
-      expect(mockRunAsync).toHaveBeenCalledWith(
-        'UPDATE UserSettings SET theme = ? WHERE id = 1',
-        ['dark'],
-      );
+      expect(mockRunAsync).toHaveBeenCalledWith('UPDATE UserSettings SET theme = ? WHERE id = 1', [
+        'dark',
+      ]);
     });
   });
